Functionality Progress
Added basis for `initial_train` method logic for real players to pick
the dominoes in their initial train.

Tested out some ascii art for displaying the dominoes in a text format,
and added functionality for displaying them in the `Domino` class, to be
used with a (future) separate object `Cursor` to draw (print) to the
screen.